+++ site/homepage/content/release-notes.md
@@ -22,6 +22,7 @@ description = "Information about changes in recent Hono releases. Includes new f
   removed.
 * The Kura protocol adapter has been removed. Eclipse Kura gateways starting with version 4 can still connect to Hono
   using Hono's standard MQTT adapter.
+* The file based device registry has been removed.
 
 ## 1.12.0