+++ site/homepage/content/release-notes.md
@@ -13,6 +13,8 @@ description = "Information about changes in recent Hono releases. Includes new f
   *hono.mongodb.connectionString* property. This has been fixed.
 * When a tenant or device gets disabled or deleted, any open AMQP or MQTT connections from clients having authenticated
   themselves as belonging to that tenant or device are getting closed now.
+* Using OpenSSL with the Quarkus based variant of Hono components did not work as described in the Secure Communication
+  guide. This has been fixed.
 
 ### API Changes