+++ site/homepage/content/release-notes.md
@@ -15,6 +15,8 @@ description = "Information about changes in recent Hono releases. Includes new f
   themselves as belonging to that tenant or device are getting closed now.
 * Using OpenSSL with the Quarkus based variant of Hono components did not work as described in the Secure Communication
   guide. This has been fixed.
+* The connection pool configuration for the HotRod client in the Quarkus variant of the Command Router component
+  didn't support using property names in camel-case. This has been fixed.
 
 ### API Changes